requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
google-generativeai>=0.3.0
openai>=1.0.0
pydantic-ai>=0.3.1
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
jinja2>=3.1.2
pydantic>=2.5.0
typer>=0.9.0
httpx>=0.25.2
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0 